+++ Rdmp.UI/SimpleDialogs/ViewSourceCodeToolTip.cs
@@ -62,7 +62,7 @@ namespace Rdmp.UI.SimpleDialogs
                     throw new FileNotFoundException("Could not find source code for file:" + Path.GetFileName(filename));
 
                 //get height of any given line
-                var coreLineHeight = e.Graphics.MeasureString("I've got a lovely bunch of coconuts" , e.Font).Height + (LINE_PADDING*2);
+                var coreLineHeight = e.Graphics.MeasureString("I've got a lovely bunch of coconuts" , e.Font).Height + (LINE_PADDING*2f);
 
                 int midpointY = HEIGHT/2;